Debug

+++ yLobby/src/br/com/battlebits/ylobby/selector/gamemode/GameModeSelector.java
@@ -119,7 +119,12 @@ public class GameModeSelector {
 				for (Entry<Integer, GameModeBase> entry : serverIds.entrySet()) {
 					GameModeBase gameModeBase = (GameModeBase) entry.getValue();
 					gameModeBase.updateOnlinePlayersOnItem();
-					selectorInventory.setItem(((Integer) entry.getKey()).intValue(), gameModeBase.getInventoryItem());
+					selectorInventory
+					.setItem(
+							entry
+							.getKey(), 
+							gameModeBase
+							.getInventoryItem());
 				}
 			}
 		};
